requirements.txt
# Core dependencies
PyQt6>=6.6.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
PyYAML>=6.0.1
Pygments>=2.16.1
pytest>=7.4.3
black>=23.10.1
pylint>=3.0.2
mypy>=1.6.1

# Code execution dependencies
docker>=6.1.3
psutil>=5.9.6

# Content processing
Markdown>=3.5.1
Jinja2>=3.1.2
python-frontmatter>=1.0.0

# Development tools
python-dotenv>=1.0.0
ipython>=8.16.1
jupyter>=1.0.0